texas_poker_app/backend/requirements.txt
fastapi==0.95.1
uvicorn[standard]
jinja2==3.1.2
websockets
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy